feat: option to disable Release Notes
Fixes #675

+++ src/extension.ts
@@ -107,7 +107,10 @@ export async function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(vscode.commands.registerCommand(WhatsNew, showWhatsNew));
   // Upon the very first initialisation create a file to indicate that the release
   // notes have been shown and not show them again.
-  if (!fs.existsSync(path.join(__dirname, 'displayReleaseNotes.txt'))) {
+  if (
+    !fs.existsSync(path.join(__dirname, 'displayReleaseNotes.txt')) &&
+    config.get<boolean>('notifications.releaseNotes')
+  ) {
     await showWhatsNew();
     fs.writeFileSync(path.join(__dirname, 'displayReleaseNotes.txt'), 'false');
   }
